requirements.txt
# Telegram Bot
aiogram==2.25.1

# Scheduler
apscheduler==3.10.4


google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0

# Google Sheets
gspread==5.12.0
oauth2client==4.1.3

# Utils
python-dotenv==1.0.0
pytz==2023.3

# FastAPI (для Mini App)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0